+++ ts/src/program/namespace/index.ts
@@ -9,6 +9,7 @@ import InstructionNamespace, { Ixs } from "./instruction";
 import TransactionNamespace, { Txs } from "./transaction";
 import RpcNamespace, { Rpcs } from "./rpc";
 import AccountNamespace, { Accounts } from "./account";
+import SimulateNamespace, { Simulate } from "./simulate";
 
 // Re-exports.
 export { State } from "./state";
@@ -16,6 +17,7 @@ export { Ixs } from "./instruction";
 export { Txs, TxFn } from "./transaction";
 export { Rpcs, RpcFn } from "./rpc";
 export { Accounts, AccountFn, ProgramAccount } from "./account";
+export { Simulate } from "./simulate";
 
 export default class NamespaceFactory {
   /**
@@ -28,12 +30,14 @@ export default class NamespaceFactory {
     coder: Coder,
     programId: PublicKey,
     provider: Provider
-  ): [Rpcs, Ixs, Txs, Accounts, State] {
+  ): [Rpcs, Ixs, Txs, Accounts, State, Simulate] {
     const idlErrors = parseIdlErrors(idl);
 
     const rpcs: Rpcs = {};
     const ixFns: Ixs = {};
     const txFns: Txs = {};
+    const simulateFns: Simulate = {};
+
     const state = StateNamespace.build(
       idl,
       coder,
@@ -46,18 +50,28 @@ export default class NamespaceFactory {
       const ix = InstructionNamespace.build(idlIx, coder, programId);
       const tx = TransactionNamespace.build(idlIx, ix);
       const rpc = RpcNamespace.build(idlIx, tx, idlErrors, provider);
+      const simulate = SimulateNamespace.build(
+        idlIx,
+        tx,
+        idlErrors,
+        provider,
+        coder,
+        programId,
+        idl
+      );
 
       const name = camelCase(idlIx.name);
 
       ixFns[name] = ix;
       txFns[name] = tx;
       rpcs[name] = rpc;
+      simulateFns[name] = simulate;
     });
 
     const accountFns = idl.accounts
       ? AccountNamespace.build(idl, coder, programId, provider)
       : {};
 
-    return [rpcs, ixFns, txFns, accountFns, state];
+    return [rpcs, ixFns, txFns, accountFns, state, simulateFns];
   }
 }